move the stack guard page between the stack and heap
where it was protecting against stack underflow into globals, which are
extremely unlikely. Far more likely, are stack overflows into the
instance's heap, where we accidentally caught those via heap guard space
instead. If an instance's `Limits` had the number of heap guard pages as
0, which is permissible, an instance with full utilization would then
map a page that had been acting as a stack guard page as read/write and
remove the guard-ness of that page.

This commit moves a page to be an outside-of-heap dedicated stack guard
page.