[lucetc] add CPU feature flags for code generation
This introduces a couple new command-line arguments and the associated programmatic APIs for
selecting optional CPU features at compile time.

Previously, we used `cranelift-native` to automatically detect and use the features available on the
host, but this can cause problems when compiling for a different CPU, even if the target triple is
constant.

Now, we can enable or disable individual features using `--target-feature=+sse3,+avx,-popcnt`, for
example. For convenience, we can also specify `--target-cpu=haswell`, for example, which enables a
fixed set of individual features.

The default behavior of auto-detection and use of features remains the same for now.