Backport - be lenient in interpreting LD
It's often more convenient to let the compiler figure out how to invoke
the linker, but setting `env LD="$CC" ... lucetc ...` poses a problem:
`$CC` can often be `"$COMPILER --option1 --option2 ..."`, e.g. `clang
--target=$TARGET`, and the current code assumes that `LD` is a single
string specifying a program, rather than some string specifying a
program along with options.

While you could require that `LD` be strictly limited to a single
string, it seems more flexible and more in line with other build systems
to interpret `LD` as space-separated tokens, the first of which
specifies the program to invoke, and the remaining tokens specifying
arguments to be passed.